+++ return_invoice_bill/doc/RELEASE_NOTES.md
@@ -1,7 +1,12 @@
 ## Module <return_invoice_bill>
 
-#### 27.01.2024
+#### 27.02.2024
 #### Version 15.0.1.0.0
 #### ADD
+- Initial commit for Automated Credit/Debit Note Generation from Return Picking     
 
-- Initial commit for Return Invoices And Bills     
+#### 21.08.2024
+#### Version 15.0.1.1.1
+#### UPDT
+- Updated the return invoice line based on the return quantity.
+- Fixed the issues while changing the language.